debugging Detail index
