+++ lib/_core/constants/move.dart
@@ -1,6 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_blog/ui/pages/auth/join_page/join_page.dart';
-import 'package:flutter_blog/ui/pages/auth/login_page/login_page.dart';
 import 'package:flutter_blog/ui/pages/post/list_page/post_list_page.dart';
 
 class Move {
@@ -12,8 +10,8 @@ class Move {
 
 Map<String, Widget Function(BuildContext)> getRouters() {
   return {
-    Move.loginPage: (context) => const LoginPage(),
-    Move.joinPage: (context) => const JoinPage(),
+    // Move.loginPage: (context) => const LoginPage(),
+    // Move.joinPage: (context) => const JoinPage(),
     Move.postListPage: (context) => PostListPage(),
   };
 }
