+++ lib/data/mock/user.dart
@@ -1,0 +1,24 @@
+class User {
+  int id;
+  String username;
+  int password;
+  String userPicUrl;
+  String location;
+  String created;
+
+  User({
+    required this.id,
+    required this.username,
+    required this.password,
+    required this.userPicUrl,
+    required this.location,
+    required this.created,
+  });
+}
+
+User user = User(id: 1, username: "ssar", password: 1234, userPicUrl: "ssar.jpg", location: "부산 서면", created: "2023-10-23");
+
+List<User> users = [
+  User(id: 1, username: "ssar", password: 1234, userPicUrl: "ssar.jpg", location: "부산 서면", created: "2023-10-23"),
+  User(id: 2, username: "cos", password: 1234, userPicUrl: "cos.jpg", location: "부산 서면", created: "2023-10-23"),
+];
