+++ lib/ui/pages/auth/login_page/widgets/login_form.dart
@@ -1,51 +1,51 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_blog/_core/constants/move.dart';
-import 'package:flutter_blog/_core/constants/size.dart';
-import 'package:flutter_blog/_core/utils/validator_util.dart';
-import 'package:flutter_blog/data/dto/user_request.dart';
-import 'package:flutter_blog/data/store/session_store.dart';
-import 'package:flutter_blog/ui/widgets/custom_auth_text_form_field.dart';
-import 'package:flutter_blog/ui/widgets/custom_elavated_button.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-
-class LoginForm extends ConsumerWidget {
-  final _formKey = GlobalKey<FormState>();
-  final _username = TextEditingController();
-  final _password = TextEditingController();
-
-  LoginForm({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    return Form(
-      key: _formKey,
-      child: Column(
-        children: [
-          CustomAuthTextFormField(
-            text: "Username",
-            obscureText: false,
-            funValidator: validateUsername(),
-            controller: _username,
-          ),
-          const SizedBox(height: mediumGap),
-          CustomAuthTextFormField(
-            text: "Password",
-            obscureText: true,
-            funValidator: validatePassword(),
-            controller: _password,
-          ),
-          const SizedBox(height: largeGap),
-          CustomElevatedButton(
-              text: "로그인",
-              funPageRoute: () {
-                if (_formKey.currentState!.validate()) {
-                  LoginReqDTO loginReqDTO = LoginReqDTO(
-                      username: _username.text, password: _password.text);
-                  ref.read(sessionProvider).login(loginReqDTO);
-                }
-              }),
-        ],
-      ),
-    );
-  }
-}
+// import 'package:flutter/material.dart';
+// import 'package:flutter_blog/_core/constants/move.dart';
+// import 'package:flutter_blog/_core/constants/size.dart';
+// import 'package:flutter_blog/_core/utils/validator_util.dart';
+// import 'package:flutter_blog/data/dto/user_request.dart';
+// import 'package:flutter_blog/data/store/session_store.dart';
+// import 'package:flutter_blog/ui/widgets/custom_auth_text_form_field.dart';
+// import 'package:flutter_blog/ui/widgets/custom_elavated_button.dart';
+// import 'package:flutter_riverpod/flutter_riverpod.dart';
+//
+// class LoginForm extends ConsumerWidget {
+//   final _formKey = GlobalKey<FormState>();
+//   final _username = TextEditingController();
+//   final _password = TextEditingController();
+//
+//   LoginForm({Key? key}) : super(key: key);
+//
+//   @override
+//   Widget build(BuildContext context, WidgetRef ref) {
+//     return Form(
+//       key: _formKey,
+//       child: Column(
+//         children: [
+//           CustomAuthTextFormField(
+//             text: "Username",
+//             obscureText: false,
+//             funValidator: validateUsername(),
+//             controller: _username,
+//           ),
+//           const SizedBox(height: mediumGap),
+//           CustomAuthTextFormField(
+//             text: "Password",
+//             obscureText: true,
+//             funValidator: validatePassword(),
+//             controller: _password,
+//           ),
+//           const SizedBox(height: largeGap),
+//           CustomElevatedButton(
+//               text: "로그인",
+//               funPageRoute: () {
+//                 if (_formKey.currentState!.validate()) {
+//                   LoginReqDTO loginReqDTO = LoginReqDTO(
+//                       username: _username.text, password: _password.text);
+//                   ref.read(sessionProvider).login(loginReqDTO);
+//                 }
+//               }),
+//         ],
+//       ),
+//     );
+//   }
+// }
