+++ lib/ui/pages/post/detail_page/post_detail_view_model.dart
@@ -1,10 +1,7 @@
-import 'package:flutter_blog/data/dto/response_dto.dart';
-import 'package:flutter_blog/data/model/post.dart';
+import 'package:flutter_blog/data/mock/post.dart';
 import 'package:flutter_blog/data/repository/post_repository.dart';
-import 'package:flutter_blog/data/store/param_store.dart';
-import 'package:flutter_blog/data/store/session_store.dart';
+import 'package:flutter_blog/main.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:logger/logger.dart';
 
 // 창고 데이터
 class PostDetailModel {
@@ -14,25 +11,17 @@ class PostDetailModel {
 
 // 창고
 class PostDetailViewModel extends StateNotifier<PostDetailModel?> {
-  PostDetailViewModel(super._state, this.ref);
-  Ref ref;
+  final mContext = navigatorKey.currentContext;
+  final Ref ref;
+  PostDetailViewModel(this.ref, super.state);
 
-  Future<void> notifyInit(int id) async {
-    Logger().d("notifyInit");
-
-    SessionUser sessionUser = ref.read(sessionProvider);
-    ResponseDTO responseDTO =
-        await PostRepository().fetchPost(sessionUser.jwt!, id);
-
-    state = PostDetailModel(responseDTO.data);
+  Future<void> notifyInit() async {
+    Post post = await PostRepository().fetchPostDetail();
+    state = PostDetailModel(post);
   }
 }
 
-// 창고 관리자
-final postDetailProvider =
-    StateNotifierProvider.autoDispose<PostDetailViewModel, PostDetailModel?>(
-        (ref) {
-  Logger().d("나실행됨? postDetailProvider");
-  int postId = ref.read(paramProvider).postDetailId!;
-  return PostDetailViewModel(null, ref)..notifyInit(postId);
+// 3. 창고 관리자
+final postDetailProvider = StateNotifierProvider.autoDispose<PostDetailViewModel, PostDetailModel?>((ref) {
+  return PostDetailViewModel(ref, null)..notifyInit();
 });
