+++ lib/ui/pages/post/list_page/post_list_view_model.dart
@@ -1,13 +1,9 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_blog/data/dto/post_request.dart';
-import 'package:flutter_blog/data/dto/response_dto.dart';
-import 'package:flutter_blog/data/model/post.dart';
+// 1. 창고 데이터
+import 'package:flutter_blog/data/mock/post.dart';
 import 'package:flutter_blog/data/repository/post_repository.dart';
-import 'package:flutter_blog/data/store/session_store.dart';
 import 'package:flutter_blog/main.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-// 1. 창고 데이터
 class PostListModel {
   List<Post> posts;
   PostListModel(this.posts);
@@ -15,44 +11,19 @@ class PostListModel {
 
 // 2. 창고
 class PostListViewModel extends StateNotifier<PostListModel?> {
-  PostListViewModel(super._state, this.ref);
-
   final mContext = navigatorKey.currentContext;
-  Ref ref;
+  final Ref ref;
+  PostListViewModel(this.ref, super.state);
 
   Future<void> notifyInit() async {
-    // jwt 가져오기
-    SessionStore sessionStore = ref.read(sessionProvider);
-
-    ResponseDTO responseDTO =
-        await PostRepository().fetchPostList(sessionStore.jwt!);
-    state = PostListModel(responseDTO.data);
-  }
-
-  Future<void> notifyAdd(PostSaveReqDTO dto) async {
-    SessionStore sessionStore = ref.read(sessionProvider);
-
-    ResponseDTO responseDTO =
-        await PostRepository().savePost(sessionStore.jwt!, dto);
-
-    if (responseDTO.code == 1) {
-      Post newPost = responseDTO.data as Post; // 1. dynamic(Post) -> 다운캐스팅
-      List<Post> newPosts = [
-        newPost,
-        ...state!.posts
-      ]; // 2. 기존 상태에 데이터 추가 [전개연산자]
-      state = PostListModel(
-          newPosts); // 3. 뷰모델(창고) 데이터 갱신이 완료 -> watch 구독자는 rebuild됨.
-      Navigator.pop(mContext!); // 4. 글쓰기 화면 pop
-    } else {
-      ScaffoldMessenger.of(mContext!).showSnackBar(
-          SnackBar(content: Text("게시물 작성 실패 : ${responseDTO.msg}")));
-    }
+    print("통신 시작");
+    List<Post> productList = await PostRepository().fetchPostList();
+    print("통신 끝 ${productList.length}");
+    state = PostListModel(posts);
   }
 }
 
-// 3. 창고 관리자 (View 빌드되기 직전에 생성됨)
-final postListProvider =
-    StateNotifierProvider.autoDispose<PostListViewModel, PostListModel?>((ref) {
-  return PostListViewModel(null, ref)..notifyInit();
+// 3. 창고 관리자
+final postListProvider = StateNotifierProvider.autoDispose<PostListViewModel, PostListModel?>((ref) {
+  return PostListViewModel(ref, null)..notifyInit();
 });
