+++ lib/ui/pages/post/list_page/wiegets/post_list_body.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_blog/data/model/post.dart';
+import 'package:flutter_blog/data/mock/post.dart';
 import 'package:flutter_blog/data/store/param_store.dart';
 import 'package:flutter_blog/ui/pages/post/detail_page/post_detail_page.dart';
 import 'package:flutter_blog/ui/pages/post/list_page/post_list_view_model.dart';
